pyproject.toml
[tool.poetry]
name = "euss-ya"
version = "0.1.0"
description = ""
authors = ["SaJH <wogur981208@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.14.0"
pillow = "^10.0.1"
django = "^4.2.5"
django-cors-headers = "^4.2.0"
django-environ = "^0.11.2"
psycopg2-binary = "^2.9.7"
gunicorn = "^21.2.0"
drf-yasg = "^1.21.7"

[tool.poetry.group.local.dependencies]
pytest = "^7.4.2"
pytest-django = "^4.5.2"

factory-boy = "^3.3.0"
faker = "^19.6.1"

django-debug-toolbar = "^4.2.0"

mypy = "^1.5.1"

django-stubs = "^4.2.4"
djangorestframework-stubs = "^3.14.2"

flake8 = "^6.1.0"
isort = "^5.12.0"
black = "^23.9.1"
pre-commit = "^3.4.0"

[tool.black]
line-length = 150
target-version = ['py311']
force-exclude = 'migrations'

[tool.isort]
profile = "black"
filter_files = true
skip_glob = ["*/migrations/*", "config/*"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
